% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishery_parameters.R
\docType{data}
\name{AdclipSelectiveFishery}
\alias{AdclipSelectiveFishery}
\title{Adipose Clip Selective Fishery with some URE and MRE}
\format{
An object of class \code{list} of length 9.
}
\usage{
AdclipSelectiveFishery
}
\description{
Adipose Clip Selective Fishery with some URE and MRE
}
\keyword{datasets}
