% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishery_simulate.R
\name{createCohort}
\alias{createCohort}
\title{Create Cohort Data Frame}
\usage{
createCohort(adclip_rate = 0.5, pns = 0.2, cohort_size = 10000L)
}
\arguments{
\item{adclip_rate}{Adipose clip rate of the cohort (related to lambda)}

\item{pns}{Proportion not legal size}

\item{cohort_size}{Size of the cohort}
}
\value{
A data frame with \code{cohort_size} rows
}
\description{
Create a data frame of random fish based on the
cohort size and the adipose clip rate
}
