% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishery_parameters.R
\docType{data}
\name{DgmPreTerminal}
\alias{DgmPreTerminal}
\title{DGM Preterminal Fishery Parameters used for the CYER report}
\format{
An object of class \code{list} of length 9.
}
\usage{
DgmPreTerminal
}
\description{
DGM Preterminal Fishery Parameters used for the CYER report
}
\keyword{datasets}
