% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishery_parameters.R
\docType{data}
\name{MixedSelectiveFishery}
\alias{MixedSelectiveFishery}
\title{Mixed Adipose Clipped Selective Fishery with higher URE and MRE}
\format{
An object of class \code{list} of length 9.
}
\usage{
MixedSelectiveFishery
}
\description{
Mixed Adipose Clipped Selective Fishery with higher URE and MRE
}
\keyword{datasets}
