% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishery_parameters.R
\docType{data}
\name{NonSelectiveFishery}
\alias{NonSelectiveFishery}
\title{Non-Selective Fishery with no disproportionate release based on adclip status}
\format{
An object of class \code{list} of length 9.
}
\usage{
NonSelectiveFishery
}
\description{
Non-Selective Fishery with no disproportionate release based on adclip status
}
\keyword{datasets}
