% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishery_simulate.R
\docType{data}
\name{OutcomeDropOff}
\alias{OutcomeDropOff}
\title{Outcome of the fishing event was Drop-Off}
\format{
An object of class \code{character} of length 1.
}
\usage{
OutcomeDropOff
}
\description{
Outcome of the fishing event was Drop-Off
}
\keyword{datasets}
