% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishery_simulate.R
\docType{data}
\name{OutcomeKept}
\alias{OutcomeKept}
\title{Outcome of the fishing event was Kept Fish}
\format{
An object of class \code{character} of length 1.
}
\usage{
OutcomeKept
}
\description{
Outcome of the fishing event was Kept Fish
}
\keyword{datasets}
