% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishery_simulate.R
\docType{data}
\name{OutcomeNoEncounter}
\alias{OutcomeNoEncounter}
\title{Outcome of the fishing event was No Encounter}
\format{
An object of class \code{character} of length 1.
}
\usage{
OutcomeNoEncounter
}
\description{
Outcome of the fishing event was No Encounter
}
\keyword{datasets}
