% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishery_simulate.R
\docType{data}
\name{OutcomeReleased}
\alias{OutcomeReleased}
\title{Outcome of the fishing event was Released fish}
\format{
An object of class \code{character} of length 1.
}
\usage{
OutcomeReleased
}
\description{
Outcome of the fishing event was Released fish
}
\keyword{datasets}
