% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishery_simulate.R
\name{selectFish}
\alias{selectFish}
\title{(Re-)Select a random cohort fish}
\usage{
selectFish(cohort_df, potential_fish)
}
\arguments{
\item{cohort_df}{Cohort data frame}

\item{potential_fish}{A random number to try as first fish}
}
\value{
Fish/row number of a fish that is encountered in a fishery
}
\description{
Select a random cohort fish, if it is dead select a different one
Do this until you find a fish or you have made as many tries as there are
fish in the cohort.
}
