% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishery_parameters.R
\name{setupFishery}
\alias{setupFishery}
\title{Setup Fishery Parameters}
\usage{
setupFishery(
  catch,
  adclip_rate,
  legal_release_mort_rate,
  nonlegal_release_mort_rate,
  legal_drop_off_mort_rate,
  nonlegal_drop_off_mort_rate,
  unclip_release_rate,
  clip_release_rate,
  prop_non_legal_size
)
}
\arguments{
\item{catch}{Total kept catch that the fishery runs until achieving}

\item{adclip_rate}{Adipose-clip rate of non-cohort fish in the fishery}

\item{legal_release_mort_rate}{Mortality rate of legal size releases}

\item{nonlegal_release_mort_rate}{Mortality rate of non-legal size releases}

\item{legal_drop_off_mort_rate}{Drop-off mortality rate of legal size catch}

\item{nonlegal_drop_off_mort_rate}{Drop-off mortality rate of non-legal size catch}

\item{unclip_release_rate}{Release rate of non-adipose-clipped fish}

\item{clip_release_rate}{Release rate of adipose-clipped fish}

\item{prop_non_legal_size}{Proportion of non-cohort fish not legal size}
}
\value{
A list with the fishery parameters for use by a fishery simulation
}
\description{
Combine the fishery parameters into a list for use by the fishery
simulation
}
