% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_methods.r
\name{sitMethod}
\alias{sitMethod}
\title{Single Index Tag Method}
\usage{
sitMethod(
  kept_mark,
  legal_release_unmark,
  kept_mark_cohort,
  escapement_mark_cohort,
  legal_release_mort_rate,
  nonlegal_release_mort_rate,
  cohort_prop_nonlegal,
  drop_off_rate,
  escapement_unmark_cohort = NULL,
  terminal_unmark_cohort = NULL
)
}
\arguments{
\item{kept_mark}{Total kept marked catch in the fishery}

\item{legal_release_unmark}{Legal released unmarked catch in the fishery}

\item{kept_mark_cohort}{Kept mark catch of the cohort (e.g. CWT estimate)}

\item{escapement_mark_cohort}{Abundance of marked cohort after the fishery}

\item{legal_release_mort_rate}{Mortality rate of legal size releases}

\item{nonlegal_release_mort_rate}{Mortality rate of non-legal size releases}

\item{cohort_prop_nonlegal}{Proportion of cohort that is not legal size}

\item{drop_off_rate}{Mortality rate of non-legal size releases}

\item{escapement_unmark_cohort}{Mortality rate of non-legal size releases}

\item{terminal_unmark_cohort}{Mortality rate of non-legal size releases}

\item{legal_release_mark}{Legal released marked catch in the fishery}

\item{kept_unmark}{Total kept unmarked catch in the fishery}
}
\value{
Fish/row number of a fish that is encountered in a fishery
}
\description{
Single Index Tag Method
}
