% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishery_simulate.R
\name{summarizeFishery}
\alias{summarizeFishery}
\title{Summarize Fishery Simulation}
\usage{
summarizeFishery(fishery_df)
}
\arguments{
\item{fishery_df}{Data frame of all fishing events from a simulation}
}
\value{
Fish/row number of a fish that is encountered in a fishery
}
\description{
Summarize the individual fishing events from a simulated fishery
}
