% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_methods.r
\name{mfaMethod}
\alias{mfaMethod}
\title{Mixed Fishery Adjustment Method}
\usage{
mfaMethod(
  kept_mark,
  legal_release_mark,
  kept_unmark,
  legal_release_unmark,
  kept_mark_cohort,
  escapement_mark_cohort,
  legal_release_mort_rate,
  nonlegal_release_mort_rate,
  cohort_prop_nonlegal,
  drop_off_rate,
  escapement_unmark_cohort = NULL,
  terminal_unmark_cohort = NULL
)
}
\arguments{
\item{kept_mark}{Total kept marked catch in the fishery}

\item{legal_release_mark}{Legal released marked catch in the fishery}

\item{kept_unmark}{Total kept unmarked catch in the fishery}

\item{legal_release_unmark}{Legal released unmarked catch in the fishery}

\item{kept_mark_cohort}{Kept mark catch of the cohort (e.g. CWT estimate)}

\item{escapement_mark_cohort}{Abundance of marked cohort after the fishery}

\item{legal_release_mort_rate}{Mortality rate of legal size releases}

\item{nonlegal_release_mort_rate}{Mortality rate of non-legal size releases}

\item{cohort_prop_nonlegal}{Proportion of cohort that is not legal size}

\item{drop_off_rate}{Drop-off and mortality rate of caught fish}

\item{escapement_unmark_cohort}{Unmarked cohort size after the fishery}

\item{terminal_unmark_cohort}{Unmarked cohort size before the fishery}
}
\value{
A data.frame with the cohort size before and after the fisher based on adclip status
}
\description{
Mixed Fishery Adjustment Method
}
