% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{sitMfaComparison}
\alias{sitMfaComparison}
\title{Run Simulation with results Compared to SIT and MFA methods}
\usage{
sitMfaComparison(
  iterations = 500,
  fishery_params = AdclipSelectiveFishery,
  cohort_pns = 0.2
)
}
\arguments{
\item{iterations}{Number of independent simulations to run}

\item{fishery_params}{Fishery parameters used for the simulations}

\item{cohort_pns}{Cohort Proportion Non-Legal Size}
}
\description{
Run Simulation with results Compared to SIT and MFA methods
}
